move unit tests to integration tests as they actually need cassandra
